card formating

+++ lib/utils.ts
@@ -16,7 +16,7 @@ export function formatDate(dateString: string): string {
   } else if (date.toDateString() === yesterday.toDateString()) {
     return 'Last Night';
   } else {
-    const options: Intl.DateTimeFormatOptions = { year: 'numeric', month: 'long', day: 'numeric' };
+    const options: Intl.DateTimeFormatOptions = { year: 'numeric', month: 'short', day: 'numeric' };
     return date.toLocaleDateString('en-US', options);
   }
 }
